feat: melhoria na organização de pastas
