+++ jogo-oito/src/main/java/chat/gpt/src/Jogo.java
@@ -5,7 +5,7 @@ public class Jogo {
     public void iniciar() {
         Tabuleiro tabuleiro = new Tabuleiro();
         Regra regra = new RegraImpl();
-        InterfaceGrafica interfaceGrafica = new InterfaceGrafica(tabuleiro, regra);
+        new InterfaceGrafica(tabuleiro, regra);
     }
 
     public void reiniciar(Tabuleiro tabuleiro) {
