+++ jogo-oito/src/main/java/chat/gpt/src/Regra.java
@@ -2,8 +2,12 @@ package chat.gpt.modelos;
 
 public interface Regra {
 
-    boolean fazerMovimento(Tabuleiro tabuleiro, Posicao poscisaoClicada);
+    boolean fazerMovimento(Tabuleiro tabuleiro, Posicao posicaoClicada);
+
     boolean verificarVitoria(Tabuleiro tabuleiro);
+
+    void reiniciarJogo(Tabuleiro tabuleiro);
+
     void embaralharPecas(Tabuleiro tabuleiro);
 
 }
