+++ jogo-oito/src/main/java/chat/gpt/src/RegraImpl.java
@@ -1,0 +1,83 @@
+package chat.gpt.modelos;
+
+import java.util.Map;
+
+public class RegraImpl implements Regra {
+
+    @Override
+    public boolean fazerMovimento(Tabuleiro tabuleiro, Posicao posicaoClicada) {
+
+        if (podeMover(tabuleiro, posicaoClicada)) {
+
+            Posicao posicaoVazia = getPosicaoVazia(tabuleiro);
+
+            if (movimentoValido(posicaoClicada, posicaoVazia)) {
+                trocarPosicao(tabuleiro, posicaoClicada, posicaoVazia);
+                return true;
+            }
+
+        }
+
+        return false;
+    }
+
+    private Posicao getPosicaoVazia(Tabuleiro tabuleiro) {
+        return tabuleiro.getTabuleiro().entrySet()
+                .stream()
+                .filter(entry -> entry.getValue().getValor() == tabuleiro.getValorPecaVazia())
+                .map(Map.Entry::getKey)
+                .findFirst()
+                .orElseThrow(RuntimeException::new);
+    }
+
+    private void trocarPosicao(Tabuleiro tabuleiro, Posicao posicaoClicada, Posicao posicaoVazia) {
+        Peca pecaVazia = new Peca(tabuleiro.getValorPecaVazia());
+
+        Peca pecaClicada = tabuleiro.getPeca(posicaoClicada);
+
+        tabuleiro.getTabuleiro().replace(posicaoClicada, pecaClicada, pecaVazia);
+        tabuleiro.getTabuleiro().replace(posicaoVazia, pecaVazia, pecaClicada);
+    }
+
+    @Override
+    public boolean verificarVitoria(Tabuleiro tabuleiro) {
+        int count = 1;
+        for (int i = 0; i < tabuleiro.getDimensao(); i++) {
+            for (int j = 0; j < tabuleiro.getDimensao(); j++) {
+                int valorPecaAtual = tabuleiro.getPeca(new Posicao(i, j)).getValor();
+
+                if (valorPecaAtual != count) return false;
+                count++;
+            }
+        }
+        return true;
+    }
+
+    @Override
+    public void reiniciarJogo(Tabuleiro tabuleiro) {
+        tabuleiro.reiniciarTabuleiro();
+    }
+
+    @Override
+    public void embaralharPecas(Tabuleiro tabuleiro) {
+        tabuleiro.embaralharPecas();
+    }
+
+    private boolean podeMover(Tabuleiro tabuleiro, Posicao posicaoClicada) {
+        return tabuleiro.getPeca(posicaoClicada).getValor() != tabuleiro.getValorPecaVazia();
+    }
+
+    private boolean movimentoValido(Posicao posicaoClicada, Posicao posicaoVazia) {
+
+        boolean mesmaLinha = posicaoVazia.getLinha() == posicaoClicada.getLinha();
+        boolean mesmaColuna = posicaoVazia.getColuna() == posicaoClicada.getColuna();
+
+        boolean movimentoCima = (posicaoVazia.getLinha() - 1) == posicaoClicada.getLinha();
+        boolean movimentoBaixo = (posicaoVazia.getLinha() + 1) == posicaoClicada.getLinha();
+        boolean movimentoEsquerda = (posicaoVazia.getColuna() - 1) == posicaoClicada.getColuna();
+        boolean movimentoDireita = (posicaoVazia.getColuna() + 1) == posicaoClicada.getColuna();
+
+        return movimentoBaixo && mesmaColuna || movimentoCima && mesmaColuna
+                || movimentoEsquerda && mesmaLinha || movimentoDireita && mesmaLinha;
+    }
+}
